A lot of changes to make the inspector live too… WIP
