+++ apps/webapp/app/presenters/v3/SpanPresenter.server.ts
@@ -6,8 +6,8 @@ import { FINAL_ATTEMPT_STATUSES, isFinalRunStatus } from "~/v3/taskStatus";
 import { BasePresenter } from "./basePresenter.server";
 
 type Result = Awaited<ReturnType<SpanPresenter["call"]>>;
-export type Span = NonNullable<NonNullable<Result>["span"]>;
-export type SpanRun = NonNullable<NonNullable<Result>["run"]>;
+export type Span = any;
+export type SpanRun = any;
 
 export class SpanPresenter extends BasePresenter {
   public async call({
@@ -23,282 +23,256 @@ export class SpanPresenter extends BasePresenter {
     spanId: string;
     runFriendlyId: string;
   }) {
-    const project = await this._replica.project.findUnique({
-      where: {
-        slug: projectSlug,
-      },
-    });
+    throw new Error("Not implemented");
 
-    if (!project) {
-      throw new Error("Project not found");
-    }
+    // const project = await this._replica.project.findUnique({
+    //   where: {
+    //     slug: projectSlug,
+    //   },
+    // });
 
-    const run = await this.getRun(spanId);
-    if (run) {
-      return {
-        type: "run" as const,
-        run,
-      };
-    }
+    // if (!project) {
+    //   throw new Error("Project not found");
+    // }
 
-    //get the run
-    const span = await this.getSpan(runFriendlyId, spanId);
+    // const run = await this.getRun(spanId);
+    // if (run) {
+    //   return {
+    //     type: "run" as const,
+    //     run,
+    //   };
+    // }
 
-    if (!span) {
-      throw new Error("Span not found");
-    }
+    // //get the run
+    // const span = await this.getSpan(runFriendlyId, spanId);
 
-    return {
-      type: "span" as const,
-      span,
-    };
+    // if (!span) {
+    //   throw new Error("Span not found");
+    // }
+
+    // return {
+    //   type: "span" as const,
+    //   span,
+    // };
   }
 
-  async getRun(spanId: string) {
-    const run = await this._replica.taskRun.findFirst({
-      select: {
-        id: true,
-        traceId: true,
-        //metadata
-        number: true,
-        taskIdentifier: true,
-        friendlyId: true,
-        isTest: true,
-        tags: {
-          select: {
-            name: true,
-          },
-        },
-        machinePreset: true,
-        lockedToVersion: {
-          select: {
-            version: true,
-            sdkVersion: true,
-          },
-        },
-        //status + duration
-        status: true,
-        startedAt: true,
-        createdAt: true,
-        updatedAt: true,
-        queuedAt: true,
-        completedAt: true,
-        logsDeletedAt: true,
-        //idempotency
-        idempotencyKey: true,
-        //delayed
-        delayUntil: true,
-        //ttl
-        ttl: true,
-        expiredAt: true,
-        //queue
-        queue: true,
-        concurrencyKey: true,
-        //schedule
-        schedule: {
-          select: {
-            friendlyId: true,
-            generatorExpression: true,
-            timezone: true,
-            generatorDescription: true,
-          },
-        },
-        //usage
-        baseCostInCents: true,
-        costInCents: true,
-        usageDurationMs: true,
-        //env
-        runtimeEnvironment: {
-          select: { id: true, slug: true, type: true },
-        },
-        payload: true,
-        payloadType: true,
-        maxAttempts: true,
-        project: {
-          include: {
-            organization: true,
-          },
-        },
-        lockedBy: {
-          select: {
-            filePath: true,
-            exportName: true,
-          },
-        },
-      },
-      where: {
-        spanId,
-      },
-    });
+  // async getRun(spanId: string) {
+  //   const run = await this._replica.taskRun.findFirst({
+  //     select: {
+  //       id: true,
+  //       traceId: true,
+  //       //metadata
+  //       number: true,
+  //       taskIdentifier: true,
+  //       friendlyId: true,
+  //       isTest: true,
+  //       tags: {
+  //         select: {
+  //           name: true,
+  //         },
+  //       },
+  //       machinePreset: true,
+  //       lockedToVersion: {
+  //         select: {
+  //           version: true,
+  //           sdkVersion: true,
+  //         },
+  //       },
+  //       //status + duration
+  //       status: true,
+  //       startedAt: true,
+  //       createdAt: true,
+  //       updatedAt: true,
+  //       queuedAt: true,
+  //       completedAt: true,
+  //       logsDeletedAt: true,
+  //       //idempotency
+  //       idempotencyKey: true,
+  //       //delayed
+  //       delayUntil: true,
+  //       //ttl
+  //       ttl: true,
+  //       expiredAt: true,
+  //       //queue
+  //       queue: true,
+  //       concurrencyKey: true,
+  //       //schedule
+  //       schedule: {
+  //         select: {
+  //           friendlyId: true,
+  //           generatorExpression: true,
+  //           timezone: true,
+  //           generatorDescription: true,
+  //         },
+  //       },
+  //       //usage
+  //       baseCostInCents: true,
+  //       costInCents: true,
+  //       usageDurationMs: true,
+  //       //env
+  //       runtimeEnvironment: {
+  //         select: { id: true, slug: true, type: true },
+  //       },
+  //       payload: true,
+  //       payloadType: true,
+  //       maxAttempts: true,
+  //       project: {
+  //         include: {
+  //           organization: true,
+  //         },
+  //       },
+  //       lockedBy: {
+  //         select: {
+  //           filePath: true,
+  //           exportName: true,
+  //         },
+  //       },
+  //     },
+  //     where: {
+  //       spanId,
+  //     },
+  //   });
 
-    if (!run) {
-      return;
-    }
+  //   if (!run) {
+  //     return;
+  //   }
 
-    const isFinished = isFinalRunStatus(run.status);
+  //   const isFinished = isFinalRunStatus(run.status);
 
-    const finishedAttempt = isFinished
-      ? await this._replica.taskRunAttempt.findFirst({
-          select: {
-            output: true,
-            outputType: true,
-            error: true,
-          },
-          where: {
-            status: { in: FINAL_ATTEMPT_STATUSES },
-            taskRunId: run.id,
-          },
-          orderBy: {
-            createdAt: "desc",
-          },
-        })
-      : null;
+  //   const finishedAttempt = isFinished
+  //     ? await this._replica.taskRunAttempt.findFirst({
+  //         select: {
+  //           output: true,
+  //           outputType: true,
+  //           error: true,
+  //         },
+  //         where: {
+  //           status: { in: FINAL_ATTEMPT_STATUSES },
+  //           taskRunId: run.id,
+  //         },
+  //         orderBy: {
+  //           createdAt: "desc",
+  //         },
+  //       })
+  //     : null;
 
-    const output =
-      finishedAttempt === null
-        ? undefined
-        : finishedAttempt.outputType === "application/store"
-        ? `/resources/packets/${run.runtimeEnvironment.id}/${finishedAttempt.output}`
-        : typeof finishedAttempt.output !== "undefined" && finishedAttempt.output !== null
-        ? await prettyPrintPacket(finishedAttempt.output, finishedAttempt.outputType ?? undefined)
-        : undefined;
+  //   const output =
+  //     finishedAttempt === null
+  //       ? undefined
+  //       : finishedAttempt.outputType === "application/store"
+  //       ? `/resources/packets/${run.runtimeEnvironment.id}/${finishedAttempt.output}`
+  //       : typeof finishedAttempt.output !== "undefined" && finishedAttempt.output !== null
+  //       ? await prettyPrintPacket(finishedAttempt.output, finishedAttempt.outputType ?? undefined)
+  //       : undefined;
 
-    const payload =
-      run.payloadType === "application/store"
-        ? `/resources/packets/${run.runtimeEnvironment.id}/${run.payload}`
-        : typeof run.payload !== "undefined" && run.payload !== null
-        ? await prettyPrintPacket(run.payload, run.payloadType ?? undefined)
-        : undefined;
+  //   const payload =
+  //     run.payloadType === "application/store"
+  //       ? `/resources/packets/${run.runtimeEnvironment.id}/${run.payload}`
+  //       : typeof run.payload !== "undefined" && run.payload !== null
+  //       ? await prettyPrintPacket(run.payload, run.payloadType ?? undefined)
+  //       : undefined;
 
-    let error: TaskRunError | undefined = undefined;
-    if (finishedAttempt?.error) {
-      const result = TaskRunError.safeParse(finishedAttempt.error);
-      if (result.success) {
-        error = result.data;
-      } else {
-        error = {
-          type: "CUSTOM_ERROR",
-          raw: JSON.stringify(finishedAttempt.error),
-        };
-      }
-    }
+  //   let error: TaskRunError | undefined = undefined;
+  //   if (finishedAttempt?.error) {
+  //     const result = TaskRunError.safeParse(finishedAttempt.error);
+  //     if (result.success) {
+  //       error = result.data;
+  //     } else {
+  //       error = {
+  //         type: "CUSTOM_ERROR",
+  //         raw: JSON.stringify(finishedAttempt.error),
+  //       };
+  //     }
+  //   }
 
-    const span = await eventRepository.getSpan(spanId, run.traceId);
+  //   const span = await eventRepository.getSpan(spanId, run.traceId);
 
-    const context = {
-      task: {
-        id: run.taskIdentifier,
-        filePath: run.lockedBy?.filePath,
-        exportName: run.lockedBy?.exportName,
-      },
-      run: {
-        id: run.friendlyId,
-        createdAt: run.createdAt,
-        tags: run.tags.map((tag) => tag.name),
-        isTest: run.isTest,
-        idempotencyKey: run.idempotencyKey ?? undefined,
-        startedAt: run.startedAt ?? run.createdAt,
-        durationMs: run.usageDurationMs,
-        costInCents: run.costInCents,
-        baseCostInCents: run.baseCostInCents,
-        maxAttempts: run.maxAttempts ?? undefined,
-        version: run.lockedToVersion?.version,
-      },
-      queue: {
-        name: run.queue,
-      },
-      environment: {
-        id: run.runtimeEnvironment.id,
-        slug: run.runtimeEnvironment.slug,
-        type: run.runtimeEnvironment.type,
-      },
-      organization: {
-        id: run.project.organization.id,
-        slug: run.project.organization.slug,
-        name: run.project.organization.title,
-      },
-      project: {
-        id: run.project.id,
-        ref: run.project.externalRef,
-        slug: run.project.slug,
-        name: run.project.name,
-      },
-      machine: run.machinePreset
-        ? machinePresetFromName(run.machinePreset as MachinePresetName)
-        : undefined,
-    };
+  //   const context = {
+  //     task: {
+  //       id: run.taskIdentifier,
+  //       filePath: run.lockedBy?.filePath,
+  //       exportName: run.lockedBy?.exportName,
+  //     },
+  //     run: {
+  //       id: run.friendlyId,
+  //       createdAt: run.createdAt,
+  //       tags: run.tags.map((tag) => tag.name),
+  //       isTest: run.isTest,
+  //       idempotencyKey: run.idempotencyKey ?? undefined,
+  //       startedAt: run.startedAt ?? run.createdAt,
+  //       durationMs: run.usageDurationMs,
+  //       costInCents: run.costInCents,
+  //       baseCostInCents: run.baseCostInCents,
+  //       maxAttempts: run.maxAttempts ?? undefined,
+  //       version: run.lockedToVersion?.version,
+  //     },
+  //     queue: {
+  //       name: run.queue,
+  //     },
+  //     environment: {
+  //       id: run.runtimeEnvironment.id,
+  //       slug: run.runtimeEnvironment.slug,
+  //       type: run.runtimeEnvironment.type,
+  //     },
+  //     organization: {
+  //       id: run.project.organization.id,
+  //       slug: run.project.organization.slug,
+  //       name: run.project.organization.title,
+  //     },
+  //     project: {
+  //       id: run.project.id,
+  //       ref: run.project.externalRef,
+  //       slug: run.project.slug,
+  //       name: run.project.name,
+  //     },
+  //     machine: run.machinePreset
+  //       ? machinePresetFromName(run.machinePreset as MachinePresetName)
+  //       : undefined,
+  //   };
 
-    return {
-      friendlyId: run.friendlyId,
-      status: run.status,
-      createdAt: run.createdAt,
-      startedAt: run.startedAt,
-      updatedAt: run.updatedAt,
-      delayUntil: run.delayUntil,
-      expiredAt: run.expiredAt,
-      completedAt: run.completedAt,
-      logsDeletedAt: run.logsDeletedAt,
-      ttl: run.ttl,
-      taskIdentifier: run.taskIdentifier,
-      version: run.lockedToVersion?.version,
-      sdkVersion: run.lockedToVersion?.sdkVersion,
-      isTest: run.isTest,
-      environmentId: run.runtimeEnvironment.id,
-      schedule: run.schedule
-        ? {
-            friendlyId: run.schedule.friendlyId,
-            generatorExpression: run.schedule.generatorExpression,
-            description: run.schedule.generatorDescription,
-            timezone: run.schedule.timezone,
-          }
-        : undefined,
-      queue: {
-        name: run.queue,
-        isCustomQueue: !run.queue.startsWith("task/"),
-        concurrencyKey: run.concurrencyKey,
-      },
-      tags: run.tags.map((tag) => tag.name),
-      baseCostInCents: run.baseCostInCents,
-      costInCents: run.costInCents,
-      totalCostInCents: run.costInCents + run.baseCostInCents,
-      usageDurationMs: run.usageDurationMs,
-      isFinished,
-      isRunning: RUNNING_STATUSES.includes(run.status),
-      payload,
-      payloadType: run.payloadType,
-      output,
-      outputType: finishedAttempt?.outputType ?? "application/json",
-      error,
-      links: span?.links,
-      context: JSON.stringify(context, null, 2),
-    };
-  }
-
-  async getSpan(runFriendlyId: string, spanId: string) {
-    const run = await this._prisma.taskRun.findFirst({
-      select: {
-        traceId: true,
-      },
-      where: {
-        friendlyId: runFriendlyId,
-      },
-    });
-
-    if (!run) {
-      return;
-    }
-
-    const span = await eventRepository.getSpan(spanId, run.traceId);
-
-    if (!span) {
-      return;
-    }
-
-    return {
-      ...span,
-      events: span.events,
-      properties: span.properties ? JSON.stringify(span.properties, null, 2) : undefined,
-      showActionBar: span.show?.actions === true,
-    };
-  }
+  //   return {
+  //     friendlyId: run.friendlyId,
+  //     status: run.status,
+  //     createdAt: run.createdAt,
+  //     startedAt: run.startedAt,
+  //     updatedAt: run.updatedAt,
+  //     delayUntil: run.delayUntil,
+  //     expiredAt: run.expiredAt,
+  //     completedAt: run.completedAt,
+  //     logsDeletedAt: run.logsDeletedAt,
+  //     ttl: run.ttl,
+  //     taskIdentifier: run.taskIdentifier,
+  //     version: run.lockedToVersion?.version,
+  //     sdkVersion: run.lockedToVersion?.sdkVersion,
+  //     isTest: run.isTest,
+  //     environmentId: run.runtimeEnvironment.id,
+  //     schedule: run.schedule
+  //       ? {
+  //           friendlyId: run.schedule.friendlyId,
+  //           generatorExpression: run.schedule.generatorExpression,
+  //           description: run.schedule.generatorDescription,
+  //           timezone: run.schedule.timezone,
+  //         }
+  //       : undefined,
+  //     queue: {
+  //       name: run.queue,
+  //       isCustomQueue: !run.queue.startsWith("task/"),
+  //       concurrencyKey: run.concurrencyKey,
+  //     },
+  //     tags: run.tags.map((tag) => tag.name),
+  //     baseCostInCents: run.baseCostInCents,
+  //     costInCents: run.costInCents,
+  //     totalCostInCents: run.costInCents + run.baseCostInCents,
+  //     usageDurationMs: run.usageDurationMs,
+  //     isFinished,
+  //     isRunning: RUNNING_STATUSES.includes(run.status),
+  //     payload,
+  //     payloadType: run.payloadType,
+  //     output,
+  //     outputType: finishedAttempt?.outputType ?? "application/json",
+  //     error,
+  //     links: span?.links,
+  //     context: JSON.stringify(context, null, 2),
+  //   };
+  // }
 }
