+++ apps/webapp/app/v3/eventRepository.server.ts
@@ -39,6 +39,8 @@ import {
   isAncestorCancelled,
   prepareEvent,
   prepareTrace,
+  rehydrateJson,
+  rehydrateShow,
 } from "~/utils/taskEvent";
 
 const MAX_FLUSH_DEPTH = 5;
@@ -395,23 +397,6 @@ export class EventRepository {
   public async getTraceSummary(traceId: string): Promise<TraceSummary | undefined> {
     return await startActiveSpan("getTraceSummary", async (span) => {
       const events = await this.readReplica.taskEvent.findMany({
-        select: {
-          id: true,
-          spanId: true,
-          parentId: true,
-          runId: true,
-          idempotencyKey: true,
-          message: true,
-          style: true,
-          startTime: true,
-          duration: true,
-          isError: true,
-          isPartial: true,
-          isCancelled: true,
-          level: true,
-          events: true,
-          environmentType: true,
-        },
         where: {
           traceId,
         },
@@ -428,24 +413,6 @@ export class EventRepository {
   public async getRunEvents(runId: string): Promise<RunPreparedEvent[]> {
     return await startActiveSpan("getRunEvents", async (span) => {
       const events = await this.readReplica.taskEvent.findMany({
-        select: {
-          id: true,
-          spanId: true,
-          parentId: true,
-          runId: true,
-          idempotencyKey: true,
-          message: true,
-          style: true,
-          startTime: true,
-          duration: true,
-          isError: true,
-          isPartial: true,
-          isCancelled: true,
-          level: true,
-          events: true,
-          environmentType: true,
-          taskSlug: true,
-        },
         where: {
           runId,
           isPartial: false,
@@ -462,221 +429,6 @@ export class EventRepository {
       }
 
       return preparedEvents;
-    });
-  }
-
-  // A Span can be cancelled if it is partial and has a parent that is cancelled
-  // And a span's duration, if it is partial and has a cancelled parent, is the time between the start of the span and the time of the cancellation event of the parent
-  public async getSpan(spanId: string, traceId: string) {
-    return await startActiveSpan("getSpan", async (s) => {
-      const spanEvent = await this.#getSpanEvent(spanId);
-
-      if (!spanEvent) {
-        return;
-      }
-
-      const preparedEvent = prepareEvent(spanEvent);
-
-      const span = await this.#createSpanFromEvent(preparedEvent);
-
-      const output = rehydrateJson(spanEvent.output);
-      const payload = rehydrateJson(spanEvent.payload);
-
-      const show = rehydrateShow(spanEvent.properties);
-
-      const properties = sanitizedAttributes(spanEvent.properties);
-
-      const messagingEvent = SpanMessagingEvent.optional().safeParse(
-        (properties as any)?.messaging
-      );
-
-      const links: SpanLink[] = [];
-
-      if (messagingEvent.success && messagingEvent.data) {
-        if (messagingEvent.data.message && "id" in messagingEvent.data.message) {
-          if (messagingEvent.data.message.id.startsWith("run_")) {
-            links.push({
-              type: "run",
-              icon: "runs",
-              title: `Run ${messagingEvent.data.message.id}`,
-              runId: messagingEvent.data.message.id,
-            });
-          }
-        }
-      }
-
-      const backLinks = spanEvent.links as any as Link[] | undefined;
-
-      if (backLinks && backLinks.length > 0) {
-        backLinks.forEach((l) => {
-          const title = String(
-            l.attributes?.[SemanticInternalAttributes.LINK_TITLE] ?? "Triggered by"
-          );
-
-          links.push({
-            type: "span",
-            icon: "trigger",
-            title,
-            traceId: l.context.traceId,
-            spanId: l.context.spanId,
-          });
-        });
-      }
-
-      const spanEvents = transformEvents(
-        preparedEvent.events,
-        spanEvent.metadata as Attributes,
-        spanEvent.environmentType === "DEVELOPMENT"
-      );
-
-      return {
-        ...spanEvent,
-        ...span.data,
-        payload,
-        output,
-        properties,
-        events: spanEvents,
-        show,
-        links,
-      };
-    });
-  }
-
-  async #createSpanFromEvent(event: PreparedEvent) {
-    return await startActiveSpan("createSpanFromEvent", async (s) => {
-      let ancestorCancelled = false;
-      let duration = event.duration;
-
-      if (!event.isCancelled && event.isPartial) {
-        await this.#walkSpanAncestors(event, (ancestorEvent, level) => {
-          if (level >= 8) {
-            return { stop: true };
-          }
-
-          if (ancestorEvent.isCancelled) {
-            ancestorCancelled = true;
-
-            // We need to get the cancellation time from the cancellation span event
-            const cancellationEvent = ancestorEvent.events.find(
-              (event) => event.name === "cancellation"
-            );
-
-            if (cancellationEvent) {
-              duration = calculateDurationFromStart(event.startTime, cancellationEvent.time);
-            }
-
-            return { stop: true };
-          }
-
-          return { stop: false };
-        });
-      }
-
-      const span = {
-        recordId: event.id,
-        id: event.spanId,
-        parentId: event.parentId ?? undefined,
-        runId: event.runId,
-        idempotencyKey: event.idempotencyKey,
-        data: {
-          message: event.message,
-          style: event.style,
-          duration,
-          isError: event.isError,
-          isPartial: ancestorCancelled ? false : event.isPartial,
-          isCancelled: event.isCancelled === true ? true : event.isPartial && ancestorCancelled,
-          startTime: getDateFromNanoseconds(event.startTime),
-          level: event.level,
-          events: event.events,
-          environmentType: event.environmentType,
-        },
-      };
-
-      return span;
-    });
-  }
-
-  async #walkSpanAncestors(
-    event: PreparedEvent,
-    callback: (event: PreparedEvent, level: number) => { stop: boolean }
-  ) {
-    const parentId = event.parentId;
-    if (!parentId) {
-      return;
-    }
-
-    await startActiveSpan("walkSpanAncestors", async (s) => {
-      let parentEvent = await this.#getSpanEvent(parentId);
-      let level = 1;
-
-      while (parentEvent) {
-        const preparedParentEvent = prepareEvent(parentEvent);
-
-        const result = callback(preparedParentEvent, level);
-
-        if (result.stop) {
-          return;
-        }
-
-        if (!preparedParentEvent.parentId) {
-          return;
-        }
-
-        parentEvent = await this.#getSpanEvent(preparedParentEvent.parentId);
-
-        level++;
-      }
-    });
-  }
-
-  async #getSpanAncestors(event: PreparedEvent, levels = 1): Promise<Array<PreparedEvent>> {
-    if (levels >= 8) {
-      return [];
-    }
-
-    if (!event.parentId) {
-      return [];
-    }
-
-    const parentEvent = await this.#getSpanEvent(event.parentId);
-
-    if (!parentEvent) {
-      return [];
-    }
-
-    const preparedParentEvent = prepareEvent(parentEvent);
-
-    if (!preparedParentEvent.parentId) {
-      return [preparedParentEvent];
-    }
-
-    const moreAncestors = await this.#getSpanAncestors(preparedParentEvent, levels + 1);
-
-    return [preparedParentEvent, ...moreAncestors];
-  }
-
-  async #getSpanEvent(spanId: string) {
-    return await startActiveSpan("getSpanEvent", async (s) => {
-      const events = await this.readReplica.taskEvent.findMany({
-        where: {
-          spanId,
-        },
-        orderBy: {
-          startTime: "asc",
-        },
-      });
-
-      let finalEvent: TaskEvent | undefined;
-
-      for (const event of events) {
-        if (event.isPartial && finalEvent) {
-          continue;
-        }
-
-        finalEvent = event;
-      }
-
-      return finalEvent;
     });
   }
 
@@ -1202,82 +954,6 @@ function isEmptyJson(json: Prisma.JsonValue) {
   return false;
 }
 
-function sanitizedAttributes(json: Prisma.JsonValue) {
-  if (json === null || json === undefined) {
-    return;
-  }
-
-  const withoutPrivateProperties = removePrivateProperties(json as Attributes);
-  if (!withoutPrivateProperties) {
-    return;
-  }
-
-  return unflattenAttributes(withoutPrivateProperties);
-}
-// removes keys that start with a $ sign. If there are no keys left, return undefined
-function removePrivateProperties(
-  attributes: Attributes | undefined | null
-): Attributes | undefined {
-  if (!attributes) {
-    return undefined;
-  }
-
-  const result: Attributes = {};
-
-  for (const [key, value] of Object.entries(attributes)) {
-    if (key.startsWith("$")) {
-      continue;
-    }
-
-    result[key] = value;
-  }
-
-  if (Object.keys(result).length === 0) {
-    return undefined;
-  }
-
-  return result;
-}
-
-function transformEvents(events: SpanEvents, properties: Attributes, isDev: boolean): SpanEvents {
-  return (events ?? []).map((event) => transformEvent(event, properties, isDev));
-}
-
-function transformEvent(event: SpanEvent, properties: Attributes, isDev: boolean): SpanEvent {
-  if (isExceptionSpanEvent(event)) {
-    return {
-      ...event,
-      properties: {
-        exception: transformException(event.properties.exception, properties, isDev),
-      },
-    };
-  }
-
-  return event;
-}
-
-function transformException(
-  exception: ExceptionEventProperties,
-  properties: Attributes,
-  isDev: boolean
-): ExceptionEventProperties {
-  const projectDirAttributeValue = properties[SemanticInternalAttributes.PROJECT_DIR];
-
-  if (projectDirAttributeValue !== undefined && typeof projectDirAttributeValue !== "string") {
-    return exception;
-  }
-
-  return {
-    ...exception,
-    stacktrace: exception.stacktrace
-      ? correctErrorStackTrace(exception.stacktrace, projectDirAttributeValue, {
-          removeFirstLine: true,
-          isDev,
-        })
-      : undefined,
-  };
-}
-
 function filteredAttributes(attributes: Attributes, prefix: string): Attributes {
   const result: Attributes = {};
 
@@ -1288,58 +964,4 @@ function filteredAttributes(attributes: Attributes, prefix: string): Attributes 
   }
 
   return result;
-}
-
-function rehydrateJson(json: Prisma.JsonValue): any {
-  if (json === null) {
-    return undefined;
-  }
-
-  if (json === NULL_SENTINEL) {
-    return null;
-  }
-
-  if (typeof json === "string") {
-    return json;
-  }
-
-  if (typeof json === "number") {
-    return json;
-  }
-
-  if (typeof json === "boolean") {
-    return json;
-  }
-
-  if (Array.isArray(json)) {
-    return json.map((item) => rehydrateJson(item));
-  }
-
-  if (typeof json === "object") {
-    return unflattenAttributes(json as Attributes);
-  }
-
-  return null;
-}
-
-function rehydrateShow(properties: Prisma.JsonValue): { actions?: boolean } | undefined {
-  if (properties === null || properties === undefined) {
-    return;
-  }
-
-  if (typeof properties !== "object") {
-    return;
-  }
-
-  if (Array.isArray(properties)) {
-    return;
-  }
-
-  const actions = properties[SemanticInternalAttributes.SHOW_ACTIONS];
-
-  if (typeof actions === "boolean") {
-    return { actions };
-  }
-
-  return;
 }
